+++ resources/assets/js/components-vue/MultiSelect.vue
@@ -292,12 +292,13 @@ export default {
             if (!this.isObject(pairs)) return false;
 
             // Check if the values are primitives, which is expected in key value pairs,
-            // also we except the first key to not be 0
+            // also we expect the first key to not be 0
             const propertyKeys = Object.getOwnPropertyNames(pairs);
+
             for (let k in propertyKeys) {
                 const key = propertyKeys[k];
 
-                if (parseInt(key) === 0) {
+                if (key === 0 || key === "0") {
                     return false;
                 }